Move comments adaptors to their own separate files
