+++ 4.EssentialApp/EssentialApp/EssentialApp/CommentsUIComposer.swift
@@ -36,36 +36,3 @@ public class CommentsUIComposer {
         return commentsViewController
     }
 }
-
-final class CommentsLoaderPresentationAdaptor: ListRefreshViewControllerDelegate {
-    private let loader: ImageCommentLoader
-    var presenter: LoadResourcePresenter<[ImageComment], CommentsViewAdapter>?
-    
-    init(loader: ImageCommentLoader) {
-        self.loader = loader
-    }
-    
-    func didStartLoadingList() {
-        presenter?.didStartLoadingResource()
-        loader.load { [weak self] result in
-            switch result {
-            case .success(let feed):
-                self?.presenter?.didCompleteLoading(with: feed)
-            case .failure(let error):
-                self?.presenter?.didCompleteLoadingResource(with: error)
-            }
-        }
-    }
-}
-
-final class CommentsViewAdapter: ResourceView {
-    private weak var controller: ListViewController?
-    
-    init(controller: ListViewController) {
-        self.controller = controller
-    }
-    
-    func display(_ viewModel: ImageCommentsViewModel) {
-        controller?.display(viewModel.comments.map(ImageCommentViewController.init(model:)))
-    }
-}
